requirements.txt
Flask==2.3.2
mysql-connector-python==8.1.0
Flask-HTTPAuth==4.7.0
werkzeug==2.3.7
